The PY3 Port: Part 8 of Infinity

+++ judge/views/blog.py
@@ -50,10 +50,10 @@ class PostList(ListView):
                 context['has_clarifications'] = clarifications.count() > 0
                 context['clarifications'] = clarifications.order_by('-date')
 
-        context['user_count'] = lazy(Profile.objects.count, int, long)
-        context['problem_count'] = lazy(Problem.objects.filter(is_public=True).count, int, long)
-        context['submission_count'] = lazy(Submission.objects.count, int, long)
-        context['language_count'] = lazy(Language.objects.count, int, long)
+        context['user_count'] = lazy(Profile.objects.count, int, int)
+        context['problem_count'] = lazy(Problem.objects.filter(is_public=True).count, int, int)
+        context['submission_count'] = lazy(Submission.objects.count, int, int)
+        context['language_count'] = lazy(Language.objects.count, int, int)
 
         context['post_comment_counts'] = {
             int(page[2:]): count for page, count in
